Implemented a work around for SpongeCommon#1238
This workaround does have a memory leak, however, this should only happen in the case where a block is marked for multiplication, but the block is broken in a way that results in no drops. This should not be of significant concern, so this shall be ignored since this is only a temporary workaround.